turma_2/2_integracao_continua_e_testes/4.testes/requirements.txt
-e .
aws_cdk.aws_s3